requirements.txt
gradio>=5.0.0,<6.0.0
Pillow>=10.0.0
requests>=2.25.0
pytest>=7.0.0
pytest-mock>=3.10.0